+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 	* Settings will be saved, if possible.
 	* The current project will be auto-saved if possible.
 	    * This will not overwrite any existing auto-saves.
+* Can select an auto-save if the editor was shutdown incorrectly last session.
 
 ## Refactors